Bug fix
Parameters were added to the parameter bucket, and binded to its current
pointer. However, when more parameters were added, the parameter bucket
array could be reallocated, invalidating the pointers the parameters were
binded to. The solution is to add every parameter to the bucket, and
then bind them to their current (and final) pointer.